+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Development
 ***********
 
 - Fix multiple Geosphere parameter and unit enums
+- Explorer: Fix wrap `(parameter, dataset)` in iterator
 
 0.92.0 (31.07.2024)
 *******************